Revert "Major changes"
This reverts commit f48aa6316e9c2508bb969643711a3a1994d26f97.

Mainly reverted use of fragments since it's confusing AF to work with and we don't really need it with the new structure/flow.

Moved "testFire()" button and function to activity_main.xml and MainActivity.kt respectively
