+++ app/build.gradle
@@ -10,6 +10,7 @@ android {
     buildToolsVersion "30.0.3"
 
     buildFeatures {
+        viewBinding true
         dataBinding true
     }
