--- app/src/main/java/com/nlm/mobilehotelappassignment/HomePage.kt
@@ -1,53 +1,0 @@
-package com.nlm.mobilehotelappassignment
-
-import android.os.Bundle
-import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
-import com.nlm.mobilehotelappassignment.databinding.FragmentHomePageBinding
-
-class HomePage : Fragment() {
-    val db = Firebase.firestore
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        val binding = DataBindingUtil.inflate<FragmentHomePageBinding>(
-            inflater,
-            R.layout.fragment_home_page, container, false
-        )
-
-        binding.button.setOnClickListener {
-            testFire()
-        }
-
-        return binding.root
-    }
-
-
-    private fun testFire() {
-        val tag = "Firestore Out"
-
-        // Create a new user with a first and last name
-        val user = hashMapOf(
-            "first" to "Ada",
-            "last" to "Lovelace",
-            "born" to 1815
-        )
-
-        // Add a new document with a generated ID
-        db.collection("users")
-            .add(user)
-            .addOnSuccessListener { documentReference ->
-                Log.d(tag, "DocumentSnapshot added with ID: ${documentReference.id}")
-            }
-            .addOnFailureListener { e ->
-                Log.w(tag, "Error adding document", e)
-            }
-    }
-}
