+++ app/src/main/java/com/nlm/mobilehotelappassignment/MainActivity.kt
@@ -2,12 +2,39 @@ package com.nlm.mobilehotelappassignment
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
+import android.view.View
+import com.google.firebase.firestore.ktx.firestore
+import com.google.firebase.ktx.Firebase
+import com.nlm.mobilehotelappassignment.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
+    val db = Firebase.firestore
+    private lateinit var binding: ActivityMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+//        setContentView(R.layout.activity_main)
+    }
 
-        setContentView(R.layout.activity_main)
+    fun testFire(view: View) {
+        val tag = "Firestore Out"
+        // Create a new user with a first and last name
+        val user = hashMapOf(
+            "first" to "Ada",
+            "last" to "Lovelace",
+            "born" to 1815
+        )
+        // Add a new document with a generated ID
+        db.collection("users")
+            .add(user)
+            .addOnSuccessListener { documentReference ->
+                Log.d(tag, "DocumentSnapshot added with ID: ${documentReference.id}")
+            }
+            .addOnFailureListener { e ->
+                Log.w(tag, "Error adding document", e)
+            }
     }
 }
