--- app/src/main/res/navigation/navigation.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/navigation"
-    app:startDestination="@id/homePage2">
-    <fragment
-        android:id="@+id/homePage2"
-        android:name="com.nlm.mobilehotelappassignment.HomePage"
-        android:label="fragment_home_page"
-        tools:layout="@layout/fragment_home_page" />
-</navigation>
