+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,3 @@
 <resources>
     <string name="app_name">Mobile Hotel App (Assignment)</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
